Quick and dirty fix for the 1500 byte limit
We don't really index the serialized JSON so
we could just mark it as such. This fixes the
current limitation though it's not ideal.

The right fix would be to rewrite the entire
backend logic to support storing the Product
object as a first class citizen (using a CRUD
API), which requires quite some changes to the
existing code.